moved into triton_pipeline
