+++ src/triton_pipeline/config/example_sequence.yaml
@@ -1,0 +1,6 @@
+/triton/pipeline_sequence_manager:
+  ros__parameters:
+    pipeline_sequence:
+    - example
+    - example
+    - example
