+++ src/triton_pipeline/launch/pipeline_launch.py
@@ -1,11 +1,18 @@
+import os
 from launch import LaunchDescription
 from launch_ros.actions import ComposableNodeContainer
 from launch_ros.actions import Node
 
+from ament_index_python.packages import get_package_share_directory
 
 def generate_launch_description():
 
     ld = LaunchDescription()
+
+    config = os.path.join(
+        get_package_share_directory('triton_pipeline'),
+        'example_sequence.yaml'
+    )
 
     pipeline_manager = Node(
         name='pipeline_manager',
@@ -13,6 +20,15 @@ def generate_launch_description():
         package='triton_pipeline',
         executable='pipeline_manager',
         output='screen'
+    )
+
+    pipeline_sequence_manager = Node(
+        name='pipeline_sequence_manager',
+        namespace='/triton',
+        package='triton_pipeline',
+        executable='pipeline_sequence_manager',
+        output='screen',
+        parameters=[config]
     )
 
     pipeline_container = ComposableNodeContainer(
@@ -23,6 +39,7 @@ def generate_launch_description():
     )
 
     ld.add_action(pipeline_manager)
+    ld.add_action(pipeline_sequence_manager)
     ld.add_action(pipeline_container)
 
     return ld
