--- src/triton_state/config/simple_sequence.yaml
@@ -1,11 +1,0 @@
-/triton/state_manager:
-  ros__parameters:
-    num_pipelines: 3
-    pipelines:
-    - example
-    - example
-    - example
-    parameters:
-    /triton/waypoint: '{x: 1, y: 2.0, z: 3, y: "hello", d: [1,2,3], e: [1.0,2.0,3.0], f:["a", "b", "c"]}'
-    - '{x: 1, y: 2, z: 3}'
-    - '{x: 1, y: 2, z: 3}'
