--- src/triton_state/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>triton_state</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="logan.fillo@gmail.com">logan</maintainer>
-  <license>TODO: License declaration</license>
-
-  <depend>triton_interfaces</depend>
-  <depend>rclpy</depend>
-  <depend>python3-yaml</depend>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
-</package>
