--- src/triton_state/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/triton_state
-[install]
-install-scripts=$base/lib/triton_state
