--- src/triton_state/setup.py
@@ -1,30 +1,0 @@
-from setuptools import setup
-from glob import glob
-import os
-
-package_name = 'triton_state'
-
-setup(
-    name=package_name,
-    version='0.0.0',
-    packages=[package_name],
-    data_files=[
-        ('share/ament_index/resource_index/packages',
-            ['resource/' + package_name]),
-        (os.path.join('share', package_name), glob('config/*.yaml')),
-        (os.path.join('share', package_name), glob('launch/*.py')),
-        ('share/' + package_name, ['package.xml']),
-    ],
-    install_requires=['setuptools'],
-    zip_safe=True,
-    maintainer='logan',
-    maintainer_email='logan.fillo@gmail.com',
-    description='TODO: Package description',
-    license='TODO: License declaration',
-    tests_require=['pytest'],
-    entry_points={
-        'console_scripts': [
-            'state_manager = triton_state.state_manager:main'
-        ],
-    },
-)
